Corrections made to CollectionRepository.kt
